Move test_all_pip_packages test and factor it
Moving test_all_pip_packages from test_pip_packages file to test_packages and extracting functions to utils so they can be reused elsewhere.
